workaround FTBFS on powerpc/gcc 4.6 in signalsloteditor
Origin: vendor
Forwarded: not-needed
Last-Update: 2011-05-07

This should workaround g++-4.6/powerpc ICE that is related to dwarf2 [1][2].
It is very likely the patch will not be needed long term as the gcc bug might
eventually be fixed.

[1] .moc/release-shared/qteditorfactory.moc:237:1: internal compiler error: in size_of_die, at dwarf2out.c:10749
Please submit a full bug report,
with preprocessed source if appropriate.

[2] https://buildd.debian.org/status/fetch.php?pkg=qt4-x11&arch=powerpc&ver=4%3A4.7.2-4&stamp=1304320067


Gbp-Pq: Name powerpc_designer_gstabs.diff