Build HTML documentation suited for offline usage
Currently, the generated documentation includes a search box which relies on an
external search service, and which does not work for local usage. Therefore, the
generated search box does not work and confuses users, so it is best to build
the documentation in offline format which does not include the search box, but
does also not confuse users.


Gbp-Pq: Name qtdoc-build-offline-docs.patch